requirements.txt
streamlit
geopandas
folium
osmnx
pandas
requests
shapely
streamlit-folium
psycopg2-binary
sqlalchemy
python-dotenv